Clean Up and update
Cleaned up the slicer and preset it to be easier to read for outsider sources.
Provided a read file on how to use it and cleaned up the unnecessary  variables in the window that popped up. Cleaned up the static class of Pokemon to Hashsets to make it faster when grabbing the info (different art).